Introduce partition protection logic, including interfaces, handlers, and additional support for secure operations. Extend dividend and voting queries with token balance and decimals to enhance transparency and functionality. Also, update `.env.sample` for expanded network configurations.

+++ contracts/contracts/layer_2/interfaces/snapshots/ISnapshots_2.sol
@@ -205,7 +205,8 @@
 
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity 0.8.18;
-
+// TODO: Remove those errors of solhint
+// solhint-disable contract-name-camelcase, var-name-mixedcase, func-name-mixedcase
 interface ISnapshots_2 {
     function ABAFAtSnapshot(
         uint256 _snapshotID
@@ -213,5 +214,6 @@ interface ISnapshots_2 {
 
     function decimalsAtSnapshot(
         uint256 _snapshotID
-    ) external view returns (uint256 decimals_);
+    ) external view returns (uint8 decimals_);
 }
+// solhint-enable contract-name-camelcase, var-name-mixedcase, func-name-mixedcase
